sv: fix topview mis-display problem
Change disp from short int to float

+++ stereo_vision.h
@@ -119,7 +119,7 @@ public:
     struct StereoData
     {
         // IMAGE -------------
-        short int disp;
+        float disp;
 
         // World coordinate system (WCS) (cm)
         int X;
